calculate gathered efficiency in sampler
